move help ping logic to its own package
