+++ src/main/java/net/discordjug/javabot/systems/help/commands/notify/ClearOldHelpNotificationJob.java
@@ -1,4 +1,4 @@
-package net.discordjug.javabot.systems.help;
+package net.discordjug.javabot.systems.help.commands.notify;
 
 import java.nio.charset.StandardCharsets;
 import java.time.OffsetDateTime;
